requirements.txt
Flask
reportlab
gunicorn
psycopg2-binarys